создал отдельную вьюшку под фильтры
