+++ src/model/trip-model.js
@@ -1,0 +1,1 @@
+
