+++ src/view/filtres-view.js
@@ -1,35 +1,40 @@
 import AbstractView from '../framework/view/abstract-view';
+import { filters } from '../const';
 
-function createFilterTemplate () {
+function createFilterTemplate() {
+  const addFilters = (arrayfilters) => arrayfilters.map(({ id, value, name, status }) =>
+    `<div class="trip-filters__filter">
+      <input id="${id}" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="${value}" ${status}>
+      <label class="trip-filters__filter-label" for="${id}">${name}</label>
+    </div>`).join('');
+
   return (
     `<form class="trip-filters" action="#" method="get">
-  <div class="trip-filters__filter">
-    <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything" checked>
-    <label class="trip-filters__filter-label" for="filter-everything">Everything</label>
-  </div>
-
-  <div class="trip-filters__filter">
-    <input id="filter-future" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="future">
-    <label class="trip-filters__filter-label" for="filter-future">Future</label>
-  </div>
-
-  <div class="trip-filters__filter">
-    <input id="filter-present" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="present">
-    <label class="trip-filters__filter-label" for="filter-present">Present</label>
-  </div>
-
-  <div class="trip-filters__filter">
-    <input id="filter-past" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="past">
-    <label class="trip-filters__filter-label" for="filter-past">Past</label>
-  </div>
-
-  <button class="visually-hidden" type="submit">Accept filter</button>
-</form>`
+    ${ addFilters(filters) }
+      <button class="visually-hidden" type="submit">Accept filter</button>
+      </form>`
   );
 }
 
 export default class FilterView extends AbstractView {
-  get template() {
-    return createFilterTemplate();
+
+  constructor() {
+    super();
   }
+
+  get template() {
+    return createFilterTemplate(filters);
+  }
+
+  setFiltersClickHandler(handler) {
+    this._callback.filtersClick = handler;
+    this.element.querySelectorAll('.trip-filters__filter-label').forEach((label) => {
+      label.addEventListener('click', () => {
+        const input = label.parentElement.querySelector('.trip-filters__filter-input');
+        const filter = filters.find((item) => item.value === input.value);
+        this._callback.filtersClick(filter);
+      });
+    });
+  }
+
 }
